+++ src/vcl-comps/devtools/Editor$/vcl.js
@@ -165,7 +165,7 @@ $([], {
 	}),
 	
     $("vcl/ui/Panel", "host", {
-        action: "toggle-component", align: "client", selected: "never",
+        action: "toggle-component", align: "client", selected: "never", executesAction: false,
         css: { 
             "background-color": "#f0f0f0", 
             "border-left": "1px solid silver",